:recycle: Move accommodation types and travel methods to travel authorizations api.
Why? I'm going to use them in the travel-authorizations store, and it makes more sense to load them from the api, than from some random component.

They should really go in a "stops-api" but I haven't needed to build one yet.
