+++ web/src/api/travel-authorizations-api.js
@@ -13,6 +13,26 @@ export const STATUSES = Object.freeze({
   SUBMITTED: "submitted",
 })
 
+// TODO: move to stops-api once that exists.
+// TODO: fetch accommodation types from backend,
+// until then, keep in sync with src/api/services/estimates/bulk-generate.ts
+export const ACCOMMODATION_TYPES = Object.freeze({
+  HOTEL: "Hotel",
+  PRIVATE: "Private",
+  OTHER: "Other:",
+})
+
+// TODO: move to stops-api once that exists.
+// TODO: load from back-end
+export const TRAVEL_METHODS = Object.freeze({
+  AIRCRAFT: "Aircraft",
+  POOL_VEHICLE: "Pool Vehicle",
+  PERSONAL_VEHICLE: "Personal Vehicle",
+  RENTAL_VEHICLE: "Rental Vehicle",
+  BUS: "Bus",
+  OTHER: "Other:",
+})
+
 export const travelAuthorizationsApi = {
   STATUSES,
   list({ where, page, perPage, ...otherParams } = {}) {
